requirements.txt
asgiref==3.2.3
certifi==2019.11.28
Django==3.0.2
django-appconf==1.0.3
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==2.2.6
django-mathfilters==0.4.0
django-select2==7.2.0
gunicorn==20.0.4
psycopg2==2.8.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
whitenoise==5.0.1
